PFP-8671 SVP: Innvilgelsesbrev sier noe om arbeidsforhold 1 som bare gjelder arbeidsforhold 2
- Oppdaterer SvpUttakResultatArbeidsforholdDto for å ta inn UttakArbeidType,
  og fikse feil der perioder som frilanser og næringsdrivende forsvant
  fra uttaklisten.
